backend/pyproject.toml
[tool.poetry]
name = "telemetry-taco-backend"
version = "0.1.0"
description = "Backend for TelemetryTaco"
authors = ["TelemetryTaco Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.14"
django = "^5.0.0"
django-ninja = "^1.0.0"
celery = "^5.3.0"
redis = "^5.0.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
django-cors-headers = "^4.3.0"
django-environ = "^0.11.0"
django-ratelimit = "^4.1.0"
django-redis = "^5.4.0"
faker = "^24.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["telemetry_taco", "core"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

